+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.4.1] - 2023-01-22
+## [0.5.1] - 2023-01-29
 
 ### CHORE
 
@@ -15,6 +15,12 @@ All notable changes to this project will be documented in this file.
 - Remove PR Lint Action
 - Action name change
 
+### DOC
+
+- Readme update
+- Update README
+- Update Changelog
+
 ### FEAT
 
 - Moved from a crate to an app
@@ -23,6 +29,12 @@ All notable changes to this project will be documented in this file.
 - Include `data` types
 - Add filename to tables and lists
 - CSV export
+- Markdown export, print summary
+- Export to default file names
+
+### FIX
+
+- CSV export fixed
 
 ### NEW